applications/model_training_api/model-training-api-requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
flasgger==0.9.7b2
python-dotenv==1.0.1
mlflow==2.14.2
scikit-learn==1.3.0
numpy==1.22.2
pandas==1.3.5
GeoAlchemy2==0.15.2
shapely==2.0.6
tensorflow== 2.13.1
geopy==2.4.1
folium==0.17.0